Update auth store tests
